+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.17] Q4 2022
+## [0.2.0] Q1 2023
+ - replaced warp by actix-web
+ - replaced BIND_ADDR and PORT by BIND_ADDR_AND_PORT
+
+## [0.1.17] Q1 2023
  - fixed GHSA-4q83-7cq4-p6wg (explicit `cargo update`)
 
-## [0.1.16] Q4 2022
+## [0.1.16] Q1 2023
  - updated dependencies in order to fix https://rustsec.org/advisories/RUSTSEC-2023-0018
  - refactored to latest iot-client-template-rs design
  - bumped to azure-iot-sdk 0.9.2